pyproject.toml
[project]
name = "news-reader-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai[google-genai,tools]>=0.152.0",
    "litellm>=1.80.5",
    "python-dotenv>=1.1.1",
]
